+++ types/src/blockchain.rs
@@ -4,11 +4,17 @@ use crate::blockchain::{
     header::BlockHeader,
 };
 
+mod height;
+
 pub mod block;
-pub mod block_height;
 pub mod consensus;
 pub mod header;
 pub mod primitives;
 
 pub type SealedBlockHeader = Sealed<BlockHeader>;
 pub type SealedBlock = Sealed<Block>;
+
+pub use height::{
+    BlockHeight,
+    DaBlockHeight,
+};
